pipeline/requirements.txt
absl-py==0.13.0
appdirs==1.4.4
astor==0.8.1
attrs==21.2.0
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.5
ConfigArgParse==1.5.2
connection-pool==0.0.3
cycler==0.10.0
dataclasses==0.8
datrie==0.8.2
decorator==4.4.2
docutils==0.17.1
filelock==3.0.12
gast==0.5.2
gitdb==4.0.7
GitPython==3.1.18
grpcio==1.40.0
h5py==3.1.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.1
ipython-genutils==0.2.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-core==4.7.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
nbformat==5.1.3
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.3.56
openslide-python==1.1.2
Pillow==8.3.2
protobuf==3.17.3
psutil==5.8.0
PuLP==2.5.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==5.4.1
ratelimiter==1.2.0.post0
requests==2.26.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
smart-open==5.2.1
smmap==4.0.0
snakemake==6.8.0
stopit==1.1.2
tabulate==0.8.9
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
threadpoolctl==2.2.0
tifffile==2020.9.3
toposort==1.6
traitlets==4.3.3
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
xgboost==1.4.2
zipp==3.5.0
